Only use the new portal implementation with -tags flatpak
